[DEV] Create pollutant.cpp/h files
