+++ src/libs/antares/study/parts/thermal/pollutant.cpp
@@ -1,0 +1,53 @@
+/*
+** Copyright 2007-2023 RTE
+** Authors: Antares_Simulator Team
+**
+** This file is part of Antares_Simulator.
+**
+** Antares_Simulator is free software: you can redistribute it and/or modify
+** it under the terms of the GNU General Public License as published by
+** the Free Software Foundation, either version 3 of the License, or
+** (at your option) any later version.
+**
+** There are special exceptions to the terms and conditions of the
+** license as they are applied to this software. View the full text of
+** the exceptions in file COPYING.txt in the directory of this software
+** distribution
+**
+** Antares_Simulator is distributed in the hope that it will be useful,
+** but WITHOUT ANY WARRANTY; without even the implied warranty of
+** MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+** GNU General Public License for more details.
+**
+** You should have received a copy of the GNU General Public License
+** along with Antares_Simulator. If not, see <http://www.gnu.org/licenses/>.
+**
+** SPDX-License-Identifier: licenceRef-GPL3_WITH_RTE-Exceptions
+*/
+
+#include "pollutant.h"
+
+namespace Antares::Data
+{
+
+void Pollutant::saveEmissionsFactor(IniFile::Section* s) const
+{
+    if (emissionFactors.empty())
+        return;
+
+    s->add("co2", emissionFactors[PollutantFactor::CO2]);
+    s->add("nh3", emissionFactors[PollutantFactor::NH3]);
+    s->add("so2", emissionFactors[PollutantFactor::SO2]);
+    s->add("nox", emissionFactors[PollutantFactor::NOX]);
+    s->add("pm2_5", emissionFactors[PollutantFactor::PM2_5]);
+    s->add("pm5", emissionFactors[PollutantFactor::PM5]);
+    s->add("pm10", emissionFactors[PollutantFactor::PM10]);
+    s->add("nmvoc", emissionFactors[PollutantFactor::NMVOC]);
+    s->add("op1", emissionFactors[PollutantFactor::OP1]);
+    s->add("op2", emissionFactors[PollutantFactor::OP2]);
+    s->add("op3", emissionFactors[PollutantFactor::OP3]);
+    s->add("op4", emissionFactors[PollutantFactor::OP4]);
+    s->add("op5", emissionFactors[PollutantFactor::OP5]);
+}
+
+} //namespace Antares::Data
